chore: rename, fix css to styles
