+++ packages/components/src/stack/hstack.tsx
@@ -1,5 +1,5 @@
 import { forwardRef } from '../system';
-import { Stack, StackProps } from './stack.component';
+import { Stack, StackProps } from './stack';
 
 export const HStack = forwardRef<StackProps, 'View'>((props, ref) => (
   <Stack {...props} flexDirection="row" ref={ref} />
