+++ packages/components/src/stack/vstack.tsx
@@ -1,5 +1,5 @@
 import { forwardRef } from '../system';
-import { Stack, StackProps } from './stack.component';
+import { Stack, StackProps } from './stack';
 
 export const VStack = forwardRef<StackProps, 'View'>((props, ref) => (
   <Stack {...props} flexDirection="column" ref={ref} />
